extract common props between summary report and combined report
