+++ src/reports/components/report-sections/summary-report-details-section.tsx
@@ -3,10 +3,10 @@
 
 import * as React from 'react';
 import { NewTabLinkWithConfirmationDialog } from 'reports/components/new-tab-link-confirmation-dialog';
-import { SummaryReportSectionProps } from 'reports/components/report-sections/summary-report-section-factory';
 import { NamedFC } from 'common/react/named-fc';
+import { BaseSummaryReportSectionProps } from 'reports/components/report-sections/base-summary-report-props';
 
-export const SummaryReportDetailsSection = NamedFC<SummaryReportSectionProps>(
+export const SummaryReportDetailsSection = NamedFC<BaseSummaryReportSectionProps>(
     'SummaryReportDetailsSection',
     props => {
         const { scanMetadata, scanTimespan, toUtcString, secondsToTimeString } = props;
